+++ src/bullets/Bullet_Basic_Missile.java
@@ -8,7 +8,7 @@ import com.jtronlabs.to_the_moon.R;
 
 public class Bullet_Basic_Missile extends Bullet_Basic{
 
-	public BulletView getBullet(Context context, Shooter shooter,double bulletSpeedY,double bulletDamage){
+	public BulletView getBullet(Context context, Shooter shooter,float bulletSpeedY,int bulletDamage){
 		
 		final int width=(int) context.getResources().getDimension(R.dimen.missile_one_width);
 		final int height=(int) context.getResources().getDimension(R.dimen.missile_one_width);
