+++ src/enemies_orbiters/Orbiter_Circle_IncSpeedOnHitView.java
@@ -6,8 +6,8 @@ import android.content.Context;
 
 public class Orbiter_Circle_IncSpeedOnHitView extends Orbiter_CircleView implements MovingViewInterface {
 	
-	public Orbiter_Circle_IncSpeedOnHitView(Context context,int score,double speedY, double speedX,double collisionDamage, 
-			double health,double probSpawnBeneficialObjecyUponDeath,
+	public Orbiter_Circle_IncSpeedOnHitView(Context context,int score,float speedY, float speedX,int collisionDamage, 
+			int health,float probSpawnBeneficialObjecyUponDeath,
 			int orbitPixelX,int orbitPixelY,int width,int height,int imageId,int circleRadius,int angularVelocityInDegrees) {
 		super( context, score, speedY,  speedX, collisionDamage, 
 				 health, probSpawnBeneficialObjecyUponDeath,
@@ -18,7 +18,7 @@ public class Orbiter_Circle_IncSpeedOnHitView extends Orbiter_CircleView impleme
 	 * increase absolute value of angular velocity by 1
 	 */
 	@Override
-	public boolean takeDamage(double amountOfDamage){
+	public boolean takeDamage(int amountOfDamage){
 		this.setAngularVelocity(this.getAngularVelocity() + this.getAngularVelocity()/(Math.abs(this.getAngularVelocity())));
 		return super.takeDamage(amountOfDamage);
 	}
