+++ src/guns/Gun_TrackingSingle.java
@@ -15,14 +15,14 @@ public  class Gun_TrackingSingle extends Gun_Tracking {
 	private Moving_ProjectileView shootTowardsMe;
 	
 	public Gun_TrackingSingle(Context context, Moving_ProjectileView shootingAtMe,
-			Shooter theShooter,Bullet bulletType,double bulletFrequency,double bulletSpeedVertical,double bulletDmg,int positionOnShooterAsAPercentage) {
+			Shooter theShooter,Bullet bulletType,float bulletFrequency,float bulletSpeedVertical,int bulletDmg,int positionOnShooterAsAPercentage) {
 		super(context,theShooter,bulletType, bulletFrequency, bulletSpeedVertical, bulletDmg, positionOnShooterAsAPercentage);
 		
 		shootTowardsMe = shootingAtMe;
 	}
 
 	public Gun_TrackingSingle(Context context,
-			Shooter theShooter,Bullet bulletType,double bulletFrequency,double bulletSpeedVertical,double bulletDmg,int positionOnShooterAsAPercentage) {
+			Shooter theShooter,Bullet bulletType,float bulletFrequency,float bulletSpeedVertical,int bulletDmg,int positionOnShooterAsAPercentage) {
 		super(context,theShooter,bulletType, bulletFrequency, bulletSpeedVertical, bulletDmg, positionOnShooterAsAPercentage);
 		
 		shootTowardsMe=null;
